pyproject.toml
[tool.poetry]
name = "nezbank"
version = "0.1.0"
description = "some description"
authors = ["selivanov"]

[tool.poetry.dependencies]
python = "^3.8"
django-extensions = "^3.1.5"
sentry-sdk = "^1.5.1"
dj-rest-auth = "^2.1.11"

[tool.poetry.dev-dependencies]
Django = "^3.2.9"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "^4.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
